+++ CHANGELOG.md
@@ -5,6 +5,25 @@ Please follow [the Keep a Changelog standard](https://keepachangelog.com/en/1.0.
 
 ## [Unreleased]
 
+## [3.4.5]
+
+### Fixed
+
+* Previously, Cadwyn was unable to handle HTTP status errors in response converters
+* Previously, Cadwyn did not set the default status code for ResponseInfo
+
+## [3.4.4]
+
+### Fixed
+
+* Request and response converters were not applied when path params were present
+
+## [3.4.3]
+
+### Added
+
+* `RouterPathParamsModifiedError` is now raised if `endpoint(...).had(path=...)` has different path params than the original route
+
 ## [3.4.2]
 
 ### Fixed